bot/requirements.txt

python-telegram-bot==21.6
httpx==0.27.2
python-json-logger==2.0.7
